BDEP on libtool.
Add a patch to avoid running configure at the end of autogen.sh as
well as again for the "proper" run with env/flags.

We aren't making any changes to autoconf input files, this is yet
another example of a github-based project who can't be bothered to run
autoconf themselves and provide a proper distribution tarball.

Written by: Stuart Henderson <sthen@openbsd.org>